requirements.txt
# Dependências do backend
crewai
supabase>=2
requests>=2.32
tenacity>=8
apscheduler>=3
python-dotenv>=1.0
pydantic>=2
psycopg[binary]>=3 